realsense_d415i/requirements.txt
asttokens==3.0.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
comm==0.2.3
ConfigArgParse==1.7.1
dash==3.2.0
decorator==5.2.1
executing==2.2.1
fastjsonschema==2.21.2
Flask==3.1.2
idna==3.10
importlib_metadata==8.7.0
ipython==9.5.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_core==5.8.1
jupyterlab_widgets==3.0.15
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
narwhals==2.5.0
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==2.2.6
open3d==0.19.0
opencv-python==4.12.0.88
packaging==25.0
parso==0.8.5
platformdirs==4.4.0
plotly==6.3.0
prompt_toolkit==3.0.52
pupil_apriltags==1.0.4.post11
pure_eval==0.2.3
Pygments==2.19.2
pyrealsense2==2.56.5.9235
pywin32==311
referencing==0.36.2
requests==2.32.5
retrying==1.4.2
rpds-py==0.27.1
stack-data==0.6.3
traitlets==5.14.3
trimesh==4.8.1
typing_extensions==4.15.0
urllib3==2.5.0
wcwidth==0.2.13
Werkzeug==3.1.3
widgetsnbextension==4.0.14
zipp==3.23.0
